Migrate ExportRecordModel to typescript
